Resolve warning: [lossy-conversions] implicit cast from long to int in compound assignment is possibly lossy It was reported by java-20-openjdk/bin/javac -deprecation

+++ src/main/java/org/monetdb/mcl/net/MapiSocket.java
@@ -1123,7 +1123,7 @@ public class MapiSocket {	/* cannot (yet) be final as nl.cwi.monetdb.mcl.net.Map
 					readPos += t;
 					readBlock();
 				} else {
-					readPos += skip;
+					readPos += (int)skip;
 					break;
 				}
 			}
